Show navigation, even on phone-sized screens.
Do not show navigation search bar on phones, since it won't fit.
Always wrap page in .container, even on root page.
This technically double renders the navigation on the root page, but the second
one which lacks the search field is displayed on top, which is still what we
want.